requirements.txt
python-telegram-bot[ext]
flask
motor
beautifulsoup4
lxml
requests
python-dotenv